# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "NewsFlow API",
        "description": "Documentação oficial da API NewsFlow",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/posts": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Lista posts com filtros opcionais por categoria, tag e título",
                "operationId": "3fda362e2e082b5357ac416e7767e435",
                "responses": {
                    "200": {
                        "description": "Posts listados com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts"
                ],
                "summary": "Cria um novo post",
                "operationId": "dcb66200c3132996151884ac232cf439",
                "responses": {
                    "201": {
                        "description": "Post criado com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                }
            }
        },
        "/api/posts/{id}": {
            "get": {
                "tags": [
                    "Posts"
                ],
                "summary": "Exibe um post específico com autor e categoria",
                "operationId": "142b5adb0520cca80cedfe09d9298332",
                "responses": {
                    "200": {
                        "description": "Post encontrado com sucesso"
                    },
                    "404": {
                        "description": "Post não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Posts"
                ],
                "summary": "Atualiza um post existente",
                "operationId": "650f4bed0846d9350903d8a17c7b503a",
                "responses": {
                    "200": {
                        "description": "Post atualizado com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão para atualizar este post"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts"
                ],
                "summary": "Remove um post",
                "operationId": "9d99f2e42af2f722a058a86245aa856f",
                "responses": {
                    "200": {
                        "description": "Post deletado com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    },
                    "403": {
                        "description": "Sem permissão para deletar este post"
                    },
                    "404": {
                        "description": "Post não encontrado"
                    }
                }
            }
        },
...
        "/posts": {
            "get": {
                "tags": [
                    "Posts Web"
                ],
                "summary": "Lista posts paginados no painel web",
                "operationId": "cde581b3119bfed4f4a0ac619622cf97",
                "responses": {
                    "200": {
                        "description": "Posts listados com sucesso"
                    }
                }
            },
            "post": {
                "tags": [
                    "Posts Web"
                ],
                "summary": "Cria um novo post via painel web",
                "operationId": "1affef85a1c5c202305fcb94bafa157e",
                "responses": {
                    "302": {
                        "description": "Post criado com sucesso"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
        "/posts/create": {
            "get": {
                "tags": [
                    "Posts Web"
                ],
                "summary": "Exibe a página de criação de post",
                "operationId": "12261540b59691ed7b6683879dda7eca",
                "responses": {
                    "200": {
                        "description": "Página de criação carregada com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        },
        "/posts/{id}/edit": {
            "get": {
                "tags": [
                    "Posts Web"
                ],
                "summary": "Exibe a página de edição de um post",
                "operationId": "4c7e3dc7fe9b1f46d97d824e6d138b7b",
                "responses": {
                    "200": {
                        "description": "Página de edição carregada com sucesso"
                    },
                    "403": {
                        "description": "Sem permissão"
                    },
                    "404": {
                        "description": "Post não encontrado"
                    }
                }
            }
        },
        "/posts/{id}": {
            "put": {
                "tags": [
                    "Posts Web"
                ],
                "summary": "Atualiza um post existente via painel web",
                "operationId": "e11abff47cfc4ec2029a65279ab9808f",
                "responses": {
                    "302": {
                        "description": "Post atualizado com sucesso"
                    },
                    "403": {
                        "description": "Sem permissão"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Posts Web"
                ],
                "summary": "Remove um post via painel web",
                "operationId": "9f9e260077be9abbec62130c26817dea",
                "responses": {
                    "302": {
                        "description": "Post deletado com sucesso"
                    },
                    "403": {
                        "description": "Sem permissão"
                    },
                    "404": {
                        "description": "Post não encontrado"
                    }
                }
            }
        },
        "/profile": {
            "get": {
                "tags": [
                    "Perfil"
                ],
                "summary": "Exibe a página de edição de perfil",
                "operationId": "8cfcc99d8d1aa366314cd8abf9cd9c9a",
                "responses": {
                    "200": {
                        "description": "Página de perfil carregada com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Perfil"
                ],
                "summary": "Deleta a conta do usuário autenticado",
                "operationId": "1d6a5c6489e35a763aa97ebef12863ce",
                "responses": {
                    "302": {
                        "description": "Conta deletada com sucesso"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Perfil"
                ],
                "summary": "Atualiza os dados do perfil do usuário autenticado",
                "operationId": "354f5b23296bed4036564563e7423f43",
                "responses": {
                    "302": {
                        "description": "Perfil atualizado com sucesso"
                    },
                    "422": {
                        "description": "Dados inválidos"
                    },
                    "401": {
                        "description": "Não autenticado"
                    }
                }
            }
        }
    },
...
}